fix: Check if recipe author is current user before delete.

+++ backend/src/use_cases/recipes/delete_recipe_use_case.py
@@ -1,16 +1,19 @@
 from uuid import UUID
 
-from fastapi import Depends, HTTPException
+from fastapi import Depends, HTTPException, status
 
 from ...repositories.recipe_repository import RecipeRepository
 
 
 class DeleteRecipeUseCase:
-    def __init__(self, repository: RecipeRepository = Depends(RecipeRepository)):
+    def __init__(self,
+                 repository: RecipeRepository = Depends(RecipeRepository),
+                 ) -> None:
         self._repository = repository
 
     def execute(
             self,
+            token: dict[str, str],
             recipe_id: UUID = None,
             recipe_title: str = None
     ) -> None:
@@ -20,8 +23,17 @@ class DeleteRecipeUseCase:
             recipe = self._repository.get_by_title(recipe_title)
         else:
             raise HTTPException(
-                status_code=404,
+                status_code=status.HTTP_404_NOT_FOUND,
                 detail="Recipe not found"
+            )
+
+        RECIPE_AUTHOR = recipe.user_id
+        CURRENT_USER_ID = token["user_id"]
+
+        if RECIPE_AUTHOR != CURRENT_USER_ID:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="You are not authorized to delete this recipe."
             )
 
         self._repository.delete(recipe)
